c-f_agents/pyproject.toml
[project]
name = "c-f-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.19",
    "pymupdf>=1.26.1",
    "selenium>=4.34.2",
]
